Completed forest fire model
Can now include different variable studies.

Data prep adjusted to fit forest fire model. The map is not triangular. the way the data prep file in Software/ML-emulation/ is now does not work with the snowmelt model . The data prep file in Transfer to Robert works for the Snowmelt model.